Move get_events and stream_events to app entity
