+++ src/snowflake/cli/_plugins/nativeapp/manager.py
@@ -14,12 +14,10 @@
 
 from __future__ import annotations
 
-import time
 from abc import ABC, abstractmethod
 from datetime import datetime
 from functools import cached_property
 from pathlib import Path
-from textwrap import dedent
 from typing import Generator, List, Optional
 
 from snowflake.cli._plugins.connection.util import make_snowsight_url
@@ -33,9 +31,6 @@ from snowflake.cli._plugins.nativeapp.entities.application import (
 from snowflake.cli._plugins.nativeapp.entities.application_package import (
     ApplicationPackageEntity,
 )
-from snowflake.cli._plugins.nativeapp.exceptions import (
-    NoEventTableForAccount,
-)
 from snowflake.cli._plugins.nativeapp.policy import AllowAlwaysPolicy, PolicyBase
 from snowflake.cli._plugins.nativeapp.project_model import (
     NativeAppProjectModel,
@@ -46,7 +41,6 @@ from snowflake.cli._plugins.stage.diff import (
 from snowflake.cli.api.console import cli_console as cc
 from snowflake.cli.api.entities.utils import (
     execute_post_deploy_hooks,
-    generic_sql_error_handler,
     sync_deploy_root_with_stage,
 )
 from snowflake.cli.api.project.schemas.entities.common import PostDeployHook
@@ -54,10 +48,8 @@ from snowflake.cli.api.project.schemas.v1.native_app.native_app import NativeApp
 from snowflake.cli.api.project.schemas.v1.native_app.path_mapping import PathMapping
 from snowflake.cli.api.project.util import (
     identifier_for_url,
-    unquote_identifier,
 )
 from snowflake.cli.api.sql_execution import SqlExecutionMixin
-from snowflake.connector import DictCursor, ProgrammingError
 
 
 class NativeAppCommandProcessor(ABC):
@@ -182,9 +174,7 @@ class NativeAppManager(SqlExecutionMixin):
 
     @cached_property
     def account_event_table(self) -> str:
-        query = "show parameters like 'event_table' in account"
-        results = self._execute_query(query, cursor_class=DictCursor)
-        return next((r["value"] for r in results if r["key"] == "EVENT_TABLE"), "")
+        return ApplicationEntity.get_account_event_table()
 
     def verify_project_distribution(
         self, expected_distribution: Optional[str] = None
@@ -370,7 +360,7 @@ class NativeAppManager(SqlExecutionMixin):
             deploy_to_scratch_stage_fn=self.deploy_to_scratch_stage_fn,
         )
 
-    def get_events(  # type: ignore [return]
+    def get_events(  # type: ignore
         self,
         since: str | datetime | None = None,
         until: str | datetime | None = None,
@@ -382,87 +372,19 @@ class NativeAppManager(SqlExecutionMixin):
         first: int = -1,
         last: int = -1,
     ) -> list[dict]:
-        record_types = record_types or []
-        scopes = scopes or []
-
-        if first >= 0 and last >= 0:
-            raise ValueError("first and last cannot be used together")
-
-        if not self.account_event_table:
-            raise NoEventTableForAccount()
-
-        # resource_attributes uses the unquoted/uppercase app and package name
-        app_name = unquote_identifier(self.app_name)
-        package_name = unquote_identifier(self.package_name)
-        org_name = unquote_identifier(consumer_org)
-        account_name = unquote_identifier(consumer_account)
-
-        # Filter on record attributes
-        if consumer_org and consumer_account:
-            # Look for events shared from a consumer account
-            app_clause = (
-                f"resource_attributes:\"snow.application.package.name\" = '{package_name}' "
-                f"and resource_attributes:\"snow.application.consumer.organization\" = '{org_name}' "
-                f"and resource_attributes:\"snow.application.consumer.name\" = '{account_name}'"
-            )
-            if consumer_app_hash:
-                # If the user has specified a hash of a specific app installation
-                # in the consumer account, filter events to that installation only
-                app_clause += f" and resource_attributes:\"snow.database.hash\" = '{consumer_app_hash.lower()}'"
-        else:
-            # Otherwise look for events from an app installed in the same account as the package
-            app_clause = f"resource_attributes:\"snow.database.name\" = '{app_name}'"
-
-        # Filter on event time
-        if isinstance(since, datetime):
-            since_clause = f"and timestamp >= '{since}'"
-        elif isinstance(since, str) and since:
-            since_clause = f"and timestamp >= sysdate() - interval '{since}'"
-        else:
-            since_clause = ""
-        if isinstance(until, datetime):
-            until_clause = f"and timestamp <= '{until}'"
-        elif isinstance(until, str) and until:
-            until_clause = f"and timestamp <= sysdate() - interval '{until}'"
-        else:
-            until_clause = ""
-
-        # Filter on event type (log, span, span_event)
-        type_in_values = ",".join(f"'{v}'" for v in record_types)
-        types_clause = (
-            f"and record_type in ({type_in_values})" if type_in_values else ""
+        return ApplicationEntity.get_events(
+            app_name=self.app_name,
+            package_name=self.package_name,
+            since=since,
+            until=until,
+            record_types=record_types,
+            scopes=scopes,
+            consumer_org=consumer_org,
+            consumer_account=consumer_account,
+            consumer_app_hash=consumer_app_hash,
+            first=first,
+            last=last,
         )
-
-        # Filter on event scope (e.g. the logger name)
-        scope_in_values = ",".join(f"'{v}'" for v in scopes)
-        scopes_clause = (
-            f"and scope:name in ({scope_in_values})" if scope_in_values else ""
-        )
-
-        # Limit event count
-        first_clause = f"limit {first}" if first >= 0 else ""
-        last_clause = f"limit {last}" if last >= 0 else ""
-
-        query = dedent(
-            f"""\
-            select * from (
-                select timestamp, value::varchar value
-                from {self.account_event_table}
-                where ({app_clause})
-                {since_clause}
-                {until_clause}
-                {types_clause}
-                {scopes_clause}
-                order by timestamp desc
-                {last_clause}
-            ) order by timestamp asc
-            {first_clause}
-            """
-        )
-        try:
-            return self._execute_query(query, cursor_class=DictCursor).fetchall()
-        except ProgrammingError as err:
-            generic_sql_error_handler(err)
 
     def stream_events(
         self,
@@ -475,62 +397,15 @@ class NativeAppManager(SqlExecutionMixin):
         consumer_app_hash: str = "",
         last: int = -1,
     ) -> Generator[dict, None, None]:
-        try:
-            events = self.get_events(
-                since=since,
-                record_types=record_types,
-                scopes=scopes,
-                consumer_org=consumer_org,
-                consumer_account=consumer_account,
-                consumer_app_hash=consumer_app_hash,
-                last=last,
-            )
-            yield from events  # Yield the initial batch of events
-            last_event_time = events[-1]["TIMESTAMP"] if events else None
-
-            while True:  # Then infinite poll for new events
-                time.sleep(interval_seconds)
-                previous_events = events
-                events = self.get_events(
-                    since=last_event_time,
-                    record_types=record_types,
-                    scopes=scopes,
-                    consumer_org=consumer_org,
-                    consumer_account=consumer_account,
-                    consumer_app_hash=consumer_app_hash,
-                )
-                if not events:
-                    continue
-
-                yield from _new_events_only(previous_events, events)
-                last_event_time = events[-1]["TIMESTAMP"]
-        except KeyboardInterrupt:
-            return
-
-
-def _new_events_only(previous_events: list[dict], new_events: list[dict]) -> list[dict]:
-    # The timestamp that overlaps between both sets of events
-    overlap_time = new_events[0]["TIMESTAMP"]
-
-    # Remove all the events from the new result set
-    # if they were already printed. We iterate and remove
-    # instead of filtering in order to handle duplicates
-    # (i.e. if an event is present 3 times in new_events
-    # but only once in previous_events, it should still
-    # appear twice in new_events at the end
-    new_events = new_events.copy()
-    for event in reversed(previous_events):
-        if event["TIMESTAMP"] < overlap_time:
-            break
-        # No need to handle ValueError here since we know
-        # that events that pass the above if check will
-        # either be in both lists or in new_events only
-        new_events.remove(event)
-    return new_events
-
-
-def _validation_item_to_str(item: dict[str, str | int]):
-    s = item["message"]
-    if item["errorCode"]:
-        s = f"{s} (error code {item['errorCode']})"
-    return s
+        return ApplicationEntity.stream_events(
+            app_name=self.app_name,
+            package_name=self.package_name,
+            interval_seconds=interval_seconds,
+            since=since,
+            record_types=record_types,
+            scopes=scopes,
+            consumer_org=consumer_org,
+            consumer_account=consumer_account,
+            consumer_app_hash=consumer_app_hash,
+            last=last,
+        )
